CPR-305 have one base class for web tests and another for messaging tests
