+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/message/listeners/OffenderDomainEventsListenerIntTest.kt
@@ -12,7 +12,7 @@ import org.mockito.kotlin.verify
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.DeliusOffenderDetail
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.Identifiers
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.Name
-import uk.gov.justice.digital.hmpps.personrecord.integration.IntegrationTestBase
+import uk.gov.justice.digital.hmpps.personrecord.integration.MultiNodeTestBase
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonEntity
 import uk.gov.justice.digital.hmpps.personrecord.message.processors.delius.NEW_OFFENDER_CREATED
 import uk.gov.justice.digital.hmpps.personrecord.message.processors.nomis.PRISONER_CREATED
@@ -31,7 +31,7 @@ import uk.gov.justice.hmpps.sqs.countAllMessagesOnQueue
 import java.time.Duration
 import java.util.concurrent.TimeUnit.SECONDS
 
-class OffenderDomainEventsListenerIntTest : IntegrationTestBase() {
+class OffenderDomainEventsListenerIntTest : MultiNodeTestBase() {
 
   @Test
   fun `should receive the message successfully when new offender event published`() {
@@ -200,11 +200,11 @@ class OffenderDomainEventsListenerIntTest : IntegrationTestBase() {
     checkTelemetry(DELIUS_RECORD_CREATION_RECEIVED, mapOf("CRN" to crn))
 
     await.atMost(Duration.ofSeconds(5)) untilCallTo {
-      cprCourtCaseEventsQueue?.sqsClient?.countAllMessagesOnQueue(cprCourtCaseEventsQueue!!.queueUrl)?.get()
+      offenderEventsQueue?.sqsClient?.countAllMessagesOnQueue(offenderEventsQueue!!.queueUrl)?.get()
     } matches { it == 0 }
 
     await.atMost(Duration.ofSeconds(5)) untilCallTo {
-      cprCourtCaseEventsQueue?.sqsDlqClient?.countAllMessagesOnQueue(cprCourtCaseEventsQueue!!.dlqUrl!!)?.get()
+      offenderEventsQueue?.sqsDlqClient?.countAllMessagesOnQueue(offenderEventsQueue!!.dlqUrl!!)?.get()
     } matches { it == 0 }
   }
 }
